backend/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis 및 캐싱
redis==5.0.1
hiredis==2.2.3

# Celery (비동기 작업 큐)
celery==5.3.4
flower==2.0.1

# AI/ML 라이브러리
torch
torchvision
transformers==4.35.2
diffusers==0.24.0
accelerate==0.25.0
safetensors==0.4.1
compel==2.0.2

# OpenAI
openai==1.3.7

# 이미지 처리
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
scipy==1.11.4

# 배경 제거 및 세그멘테이션
rembg==2.0.53
segment-anything==1.0

# 유틸리티
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# HTTP 클라이언트
httpx==0.25.2
aiohttp==3.9.1

# 로깅 및 모니터링
loguru==0.7.2
prometheus-client==0.19.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 개발 도구
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# CORS
python-cors==1.0.0

# 파일 업로드
aiofiles==23.2.1

# 이미지 최적화
imageio==2.33.0
scikit-image==0.22.0

# ControlNet 및 고급 이미지 생성
controlnet-aux==0.0.7
# xformers==0.0.23  # GPU 메모리 최적화

# 시즈널 스토리 생성 (날씨, 트렌드 등)
requests==2.31.0  # 날씨 API 호출
pytz==2023.3  # 시간대 처리
beautifulsoup4==4.12.2  # 트렌드 크롤링
lxml==4.9.3  # HTML 파싱
pytrends==4.9.2  # Google Trends API

# MySQL 데이터베이스
pymysql==1.1.0  # MySQL 연결
cryptography==41.0.7  # PyMySQL 암호화
